requested changes

+++ gateway/managers/syncman/syncman_crud.go
@@ -356,16 +356,20 @@ func (s *Manager) DeleteCollectionRules(ctx context.Context, project, dbAlias, c
 		return http.StatusBadRequest, err
 	}
 
-	resourceID := config.GenerateResourceID(s.clusterID, project, config.ResourceDatabaseRule, dbAlias, col, "rule")
-	databaseRulesConfig, ok := projectConfig.DatabaseRules[resourceID]
-	if !ok {
-		return http.StatusBadRequest, helpers.Logger.LogError(helpers.GetRequestID(ctx), fmt.Sprintf("Unable to delete collection rules, provided db alias (%s) does not exists", dbAlias), nil, nil)
+	if !s.checkIfDbAliasExists(projectConfig.DatabaseConfigs, dbAlias) {
+		return http.StatusBadRequest, helpers.Logger.LogError(helpers.GetRequestID(ctx), fmt.Sprintf("Unable to delete collection rules provided db alias (%s) does not exists", dbAlias), nil, nil)
 	}
 
-	databaseRulesConfig.Rules = make(map[string]*config.Rule)
+	resourceID := config.GenerateResourceID(s.clusterID, project, config.ResourceDatabaseRule, dbAlias, col, "rule")
+	_, ok := projectConfig.DatabaseRules[resourceID]
+	if !ok {
+		return http.StatusBadRequest, helpers.Logger.LogError(helpers.GetRequestID(ctx), fmt.Sprintf("Unable to delete collection rules, provided table or collection (%s) does not exists", col), nil, nil)
+	}
+
+	delete(projectConfig.DatabaseRules, resourceID)
 
 	if err := s.modules.SetDatabaseRulesConfig(ctx, projectConfig.DatabaseRules); err != nil {
-		return http.StatusInternalServerError, helpers.Logger.LogError(helpers.GetRequestID(ctx), "Unable to set crud config", err, nil)
+		return http.StatusInternalServerError, helpers.Logger.LogError(helpers.GetRequestID(ctx), "Unable to set database rules config", err, nil)
 	}
 
 	if err := s.store.DeleteResource(ctx, resourceID); err != nil {
